improvements made but not tested !!!
nearby place are not offseted anymore.  there is 2 loops in button
click. first is checking for white listed places if there is none in 25
meter then the second part sends the request for nearby places and
looping through to find closest. nearby places kept in the variable
Venues, it must be empty on every button click.